+++ src/Engines/WebGPU/Extensions/index.ts
@@ -1,0 +1,14 @@
+import "./engine.alpha";
+import "./engine.computeShader";
+import "./engine.cubeTexture";
+import "./engine.dynamicBuffer";
+import "./engine.dynamicTexture";
+import "./engine.multiRender";
+import "./engine.query";
+import "./engine.rawTexture";
+import "./engine.readTexture";
+import "./engine.renderTarget";
+import "./engine.renderTargetCube";
+import "./engine.storageBuffer";
+import "./engine.uniformBuffer";
+import "./engine.videoTexture";
