+++ src/Engines/index.ts
@@ -6,6 +6,7 @@ export * from "./engine";
 export * from "./engineStore";
 export * from "./nullEngine";
 export * from "./Extensions/index";
+export * from "./WebGPU/Extensions/index";
 export * from "./IPipelineContext";
 export * from "./WebGL/webGLPipelineContext";
 export * from "./WebGPU/webgpuConstants";
